Move integration test to ariesvcx
