+++ agency_client/src/utils/constants.rs
@@ -1,4 +1,3 @@
-// The byte arrays come directly from the agency team's documentation
 pub const CREATE_KEYS_V2_RESPONSE: &'static [u8; 343] = &[123, 34, 109, 101, 115, 115, 97, 103, 101, 34, 58, 34, 123, 92, 34, 64, 116, 121, 112, 101, 92, 34, 58, 92, 34, 100, 105, 100, 58, 115, 111, 118, 58, 49, 50, 51, 52, 53, 54, 55, 56, 57, 97, 98, 99, 100, 101, 102, 103, 104, 105, 49, 50, 51, 52, 59, 115, 112, 101, 99, 47, 99, 111, 110, 110, 101, 99, 116, 105, 110, 103, 47, 48, 46, 54, 47, 75, 69, 89, 95, 67, 82, 69, 65, 84, 69, 68, 92, 34, 44, 92, 34, 119, 105, 116, 104, 80, 97, 105, 114, 119, 105, 115, 101, 68, 73, 68, 92, 34, 58, 92, 34, 77, 78, 101, 112, 101, 83, 87, 116, 71, 102, 104, 110, 118, 56, 106, 76, 66, 49, 115, 70, 90, 67, 92, 34, 44, 92, 34, 119, 105, 116, 104, 80, 97, 105, 114, 119, 105, 115, 101, 68, 73, 68, 86, 101, 114, 75, 101, 121, 92, 34, 58, 92, 34, 67, 55, 51, 77, 82, 110, 110, 115, 52, 113, 85, 106, 82, 53, 78, 52, 76, 82, 119, 84, 121, 105, 88, 86, 80, 75, 80, 114, 65, 53, 113, 52, 76, 67, 84, 56, 80, 90, 122, 120, 86, 100, 116, 57, 92, 34, 125, 34, 44, 34, 114, 101, 99, 105, 112, 105, 101, 110, 116, 95, 118, 101, 114, 107, 101, 121, 34, 58, 34, 50, 112, 70, 109, 113, 97, 98, 119, 75, 82, 82, 109, 80, 54, 76, 117, 67, 99, 121, 65, 70, 101, 107, 56, 77, 109, 68, 75, 107, 107, 102, 100, 72, 111, 100, 116, 57, 103, 90, 49, 67, 88, 74, 52, 34, 44, 34, 115, 101, 110, 100, 101, 114, 95, 118, 101, 114, 107, 101, 121, 34, 58, 34, 65, 66, 117, 89, 121, 84, 87, 90, 120, 113, 53, 88, 98, 105, 54, 90, 69, 119, 119, 121, 49, 97, 51, 69, 88, 51, 90, 68, 56, 100, 105, 112, 115, 109, 102, 106, 81, 53, 75, 116, 71, 116, 115, 120, 34, 125];
 pub const UPDATE_MESSAGES_RESPONSE: &'static [u8; 147] = &[129, 167, 98, 117, 110, 100, 108, 101, 100, 145, 220, 0, 119, 208, 130, 208, 165, 64, 116, 121, 112, 101, 208, 130, 208, 164, 110, 97, 109, 101, 208, 187, 77, 83, 71, 95, 83, 84, 65, 84, 85, 83, 95, 85, 80, 68, 65, 84, 69, 68, 95, 66, 89, 95, 67, 79, 78, 78, 83, 208, 163, 118, 101, 114, 208, 163, 49, 46, 48, 208, 178, 117, 112, 100, 97, 116, 101, 100, 85, 105, 100, 115, 66, 121, 67, 111, 110, 110, 115, 208, 145, 208, 130, 208, 171, 112, 97, 105, 114, 119, 105, 115, 101, 68, 73, 68, 208, 182, 55, 57, 82, 78, 82, 113, 77, 87, 52, 72, 106, 107, 97, 111, 98, 101, 72, 105, 69, 77, 110, 101, 208, 164, 117, 105, 100, 115, 208, 145, 208, 167, 111, 103, 109, 49, 121, 50, 102];
 pub const UPDATE_PROFILE_RESPONSE: &'static [u8; 57] = &[129, 167, 98, 117, 110, 100, 108, 101, 100, 145, 220, 0, 37, 208, 129, 208, 165, 64, 116, 121, 112, 101, 208, 130, 208, 164, 110, 97, 109, 101, 208, 175, 67, 79, 78, 70, 73, 71, 83, 95, 85, 80, 68, 65, 84, 69, 68, 208, 163, 118, 101, 114, 208, 163, 49, 46, 48];
@@ -23,12 +22,3 @@ pub const AGENCY_CONFIGS_UPDATED: &str = r#"
 {
 	"@type": "did:sov:123456789abcdefghi1234;spec/configs/1.0/CONFIGS_UPDATED"
 }"#;
-
-#[cfg(test)]
-pub const MY1_SEED: &str = "00000000000000000000000000000My1";
-#[cfg(test)]
-pub const MY2_SEED: &str = "00000000000000000000000000000My2";
-#[cfg(test)]
-pub const MY3_SEED: &str = "00000000000000000000000000000My3";
-#[cfg(test)]
-pub const MY4_SEED: &str = "00000000000000000000000000000My4";
